Lecture23_AgenticAI/ResearchAgent/requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
google-genai
google-search-results


# Ignore the .env file that contains your API keys  
# .env    in gitignore